test/lasergame-leaderboard-project/backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
gunicorn
python-dotenv